# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shopify MCP AI Shopping Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{storeDomain}/chat:
    post:
      operationId: sendChat
      tags: [chat]
      summary: Send a chat message (returns SSE stream)
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: SSE stream of chat events
          content:
            text/event-stream:
              schema:
                type: string
...
  /stores/{storeDomain}/feedback:
    post:
      operationId: submitFeedback
      tags: [chat]
      summary: Submit thumbs up/down feedback on an AI response
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
        - name: x-session-id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackBody"
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
components:
  schemas:
...
    ChatRequest:
      type: object
      properties:
        sessionId:
          type: string
        conversationId:
          type: ["integer", "null"]
        message:
          type: string
        imageBase64:
          type: string
          description: Base64-encoded image data URL for visual product search
        context:
          $ref: "#/components/schemas/ChatContext"
      required:
        - sessionId
        - message
    ChatContext:
      type: object
      properties:
        productHandle:
          type: string
        collectionHandle:
          type: string
        cartToken:
          type: string
        searchMode:
          type: boolean
...
    FeedbackBody:
      type: object
      properties:
        messageId:
          type: string
        rating:
          type: string
          enum: [positive, negative]
        messageContent:
          type: string
        comment:
          type: string
        conversationId:
          type: integer
      required:
        - messageId
        - rating
    FeedbackResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success